fixed some compile errors (.NET)
git-svn-id: file:///.data/svn_convert/svn@5850 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/sim/plugins/remote/SimControl.cpp
@@ -136,7 +136,7 @@ bool IPC::process(const BSTR &in_str, BSTR *out_str)
         if (s[i] == SLOT_IN)
             continue;
         if (s[i] == SLOT_OUT){
-            CComBSTR res(mem);
+            CComBSTR res((int)mem);
             *out_str = res.Copy();
         }
         UnmapViewOfFile(mem);
